file
The Problem
The company wants to modify its products based on its target customers from different customer segments. Analyzing which customer segment is most likely to buy the product is required to meet the business goals.

Description
This project entails a detailed analysis of a company's ideal customers. A dataset containing customers' information with the company was utilized. Embedded within the dataset are errors that were removed and stored in a separate file for rectification. 

This report presents a concise analysis and conclusions to better understand the company's customers and make it easier for them to modify products based on their target customers from different customer segments. Analyzing which customer segments will most likely purchase a product and then market the product only on that particular segment helps save time and resources.

The results obtained from the analysis are definitive. Graduation level customers, one of the top 20 earners, have the highest cumulative yearly income with a 45% income rate. Although there is a high rate of products purchased by customers, especially graduation level customers, there is a low campaign response rate on all educational levels and marital statuses, with the highest responsive rate of 21% for Ph.D. level customers.

Conclusion
Comparing visually, we can say that Graduation and Ph.D. education level customers generate the most income for the company, and their response rate to campaigns is relatively high. Although there is a high rate of products purchased by customers, there is a low campaign response rate on all educational levels and marital statuses, with the highest responsive rate of 21% for Ph.D. level customers. I would recommend that the company improve its campaigning strategy to improve customers' response further.

The low yearly household income of Basic education level customers reflects why they are not on the top 20 list and why their purchasing rate is meager. Hence, spending or modifying products for this customer segment will be a loss and waste of resources for the company.